[None][fix] stop the gather act-fusion test staging weights through int32
Adding the SiTU activation grew this test's grid from 144 to 216 cases, and
the extra ones are all gated, which doubles the weight tensor. That tipped
the ep_size=1 cases (256 local experts) into CUDA OOM -- including some
pre-existing SwiGLU ones that were already close to the edge.

The avoidable cost is the staging buffer: drawing into int32 and casting to
bfloat16 keeps both live at once, ~100 GB for a single ep_size=1 gated case,
for a tensor that ends up at ~34 GB. Draw straight into bfloat16 instead;
the values are small integers, so they are exactly representable and the
test's reference comparison is unchanged.

+++ tests/unittest/_torch/thop/parallel/test_cute_dsl_moe.py
@@ -1226,16 +1226,17 @@ def test_nvfp4_gather_grouped_gemm_act_fusion_blackwell(
     num_valid_permuted_tokens = total_num_padded_tokens.item()
 
     # Create input tensors (original size, not permuted)
-    a = torch.randint(-5, 5, (num_tokens, hidden_size), dtype=torch.int32, device="cuda").to(
-        torch.bfloat16
-    )
+    # Draw straight into bfloat16: the int32 staging buffer for `b` is twice
+    # the size of the tensor it produces, and at ep_size=1 the pair is ~100 GB
+    # live at once, which is what makes this grid OOM.
+    a = torch.randint(-5, 5, (num_tokens, hidden_size), dtype=torch.bfloat16, device="cuda")
     b = torch.randint(
         -5,
         5,
         (num_local_experts, interm_size * weight_n_multiplier, hidden_size),
-        dtype=torch.int32,
+        dtype=torch.bfloat16,
         device="cuda",
-    ).to(torch.bfloat16)
+    )
 
     # Quantize inputs to FP4
     a_global_sf = a.abs().max().float() / (448 * 6)
